% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{acfplot}
\alias{acfplot}
\title{Autocorrelation plot}
\usage{
acfplot(x)
}
\arguments{
\item{x}{A vector with simulations from a MCMC process (eg from a \code{BUGS}
or \code{JAGS} run)}
}
\description{
Plots the ACF function
}
\examples{
\dontrun{
}
}
\author{
Gianluca Baio
}
\keyword{Autocorrelation}
\keyword{function}
