% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{betaPar2}
\alias{betaPar2}
\title{Compute the parameters of a Beta distribution, given a prior guess for key
parameters. Based on "Bayesian ideas and data analysis", page 100.
Optimisation method to identify the values of a,b that give required
conditions on the Beta distribution}
\usage{
betaPar2(mode, upp, prob)
}
\arguments{
\item{mode}{The implied mode of the distribution}

\item{upp}{An upper bound value for the distribution}

\item{prob}{The estimated probability that theta <= upp}
}
\value{
The list of relevant output including the values for the
parameters of the Beta distribution and some underlying summary statistics
of the resulting variable
}
\description{
Compute the parameters of a Beta distribution, given a prior guess for key
parameters. Based on "Bayesian ideas and data analysis", page 100.
Optimisation method to identify the values of a,b that give required
conditions on the Beta distribution
}
\examples{
res=betaPar2(.6,.7,.9)

}
