% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{betaplot}
\alias{betaplot}
\title{Trial-and-error Beta plot}
\usage{
betaplot(a_max = 30, b_max = 30, step = 0.01)
}
\arguments{
\item{a_max}{The maximum value for the parameter \code{a} of the Beta distribution}

\item{b_max}{The maximum value for the parameter \code{b} of the Beta distribution}

\item{step}{The increment in the grid of values for \code{a} and \code{b}}
}
\description{
Provides a quick and dirty, trial-and-error tool to identify suitable values for the the
parameters of a Beta distribution to match set properties (eg mean, sd, 95\% interval)
}
\examples{
\dontrun{
betaplot()
}
}
\author{
Gianluca Baio
}
\keyword{Beta}
\keyword{distribution}
