% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{change_of_variable}
\alias{change_of_variable}
\title{Change of variable}
\usage{
change_of_variable(f, g, p, args.p = NULL, y)
}
\arguments{
\item{f}{an expression with the function y=f(x)}

\item{g}{an expression with the inverse function x=g(y)}

\item{p}{a function for the distribution for p_X(x). Can be specified as a
built-in, named function (e.g. \code{dnorm}), or defined by the user, e.g.
\code{p=function(x,mean,sd){(1/(sqrt(2*pi)*sd))*exp(-0.5*((x-mean)/sd)^2)}}.
Note that the main argument to the function \emph{must} be named \code{x}}

\item{args.p}{a list with optional parameters for the function described in
the expression \code{p}}

\item{y}{a vector of values for the variable \code{y}}
}
\value{
The value of the density p_Y(y) computed using the change of
variable
}
\description{
Constructs a density using the change of variable rule
}
\examples{
f=expression(exp(x)/(1+exp(x)))
g=expression(log(y/(1-y)))
p=dnorm
args.p=list(mean=-.405,sd=.413)
y=seq(0,1,.001)
change_of_variable(f,g,p,args.p,y)
tibble(x=y,y=change_of_variable(f,g,p,args.p,y)) |>
ggplot(aes(x,y)) + geom_line()
ggplot() + stat_function(
fun=change_of_variable, args=list(f=f,g=g,p=p,args.p=args.p),n=1001
)

}
\author{
Gianluca Baio
}
