% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gammaPar}
\alias{gammaPar}
\title{Computes the parameters of a Gamma distribution so that the mean and
standard dev are the input (m,s)}
\usage{
gammaPar(m, s)
}
\arguments{
\item{m}{The implied mean for the underlying Beta distribution}

\item{s}{The implied standard deviation for the underlying Beta distribution}
}
\value{
The list of relevant output including the values for the
parameters of the Beta distribution (alpha and beta)
}
\description{
Computes the parameters of a Gamma distribution so that the mean and
standard dev are the input (m,s)
}
\examples{
gammaPar(12,3)

}
