% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{gammaplot}
\alias{gammaplot}
\title{Trial-and-error Gamma plot}
\usage{
gammaplot(shape_max = 30, rate_max = 30, step = 0.01)
}
\arguments{
\item{shape_max}{The maximum value for the parameter \code{shape} of the Gamma distribution}

\item{rate_max}{The maximum value for the parameter \code{rate} of the Gamma distribution}

\item{step}{The increment in the grid of values for \code{shape} and \code{rate}}
}
\description{
Provides a quick and dirty, trial-and-error tool to identify suitable values for the the
parameters of a Gamma distribution to match set properties (eg mean, sd, 95\% interval)
}
\examples{
\dontrun{
gammaplot()
}
}
\author{
Gianluca Baio
}
\keyword{Gamma}
\keyword{distribution}
