% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmhe.R
\name{logit}
\alias{logit}
\title{Computes the logit of a number}
\usage{
logit(x)
}
\arguments{
\item{x}{a number between 0 and 1}
}
\value{
logit(x)
}
\description{
Computes the logit of a number
}
\examples{
logit(.2)

}
