% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{logitPar}
\alias{logitPar}
\title{Computes the parameters of a Normal distribution \emph{on the logit scale},
so that, \emph{on the natural scale}, the range where most of the mass is
included is between the input 'low' and 'upp'}
\usage{
logitPar(low, upp)
}
\arguments{
\item{low}{The lower extreme of an implied range that is supposed to cover
"most" of the mass under the natural scale of the distribution of the
parameter (defined in the interval 0--1)}

\item{upp}{The upper extreme of an implied range that is supposed to cover
"most" of the mass under the natural scale of the distribution of the
parameter (defined in the interval 0--1)}
}
\value{
The list of relevant output including the values for the
parameters of the normal distribution (mulogit and sigmalogit),
\strong{on the logit scale}
}
\description{
Computes the parameters of a Normal distribution \emph{on the logit scale},
so that, \emph{on the natural scale}, the range where most of the mass is
included is between the input 'low' and 'upp'
}
\examples{
logitPar(0.04,0.12)

}
