% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmhe.R
\name{lognPar}
\alias{lognPar}
\title{Computes mean and variance of a logNormal distribution
so that the parameters on the natural scale are mu and sigma}
\usage{
lognPar(m, s)
}
\arguments{
\item{m}{The implied mean for the underlying Beta distribution}

\item{s}{The implied standard deviation for the underlying Beta distribution}
}
\value{
The list of relevant output including the values for the
parameters of the logNormal distribution in terms of the mean on the log
scale (mulog) and the sd on the log scale (sigmalog)
}
\description{
Computes mean and variance of a logNormal distribution
so that the parameters on the natural scale are mu and sigma
}
\examples{
lognPar(3,.15)
}
