% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmhe.R
\name{mytraceplot}
\alias{mytraceplot}
\title{Makes a traceplot (eg to visualise MCMC simulations from multiple chains)}
\usage{
mytraceplot(node, model = m, title = "", lab = "")
}
\arguments{
\item{node}{a \emph{string} with the name of the node to be plotted, eg "theta"
(in quotes)}

\item{model}{the name of the object containing the MCMC simulations}

\item{title}{the title of the graph (defaults to nothing)}

\item{lab}{the label to write on the y-axis (defaults to nothing)}
}
\value{
the graph with the traceplot
}
\description{
Makes a traceplot (eg to visualise MCMC simulations from multiple chains)
}
\examples{

}
