% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmhe.R
\name{odds2probs}
\alias{odds2probs}
\title{Maps from odds to probabilities}
\usage{
odds2probs(odds)
}
\arguments{
\item{odds}{the odds (ratio of p/(1-p))}
}
\value{
the value p
}
\description{
Maps from odds to probabilities
}
\examples{
odds2probs(4)

}
