% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{OR}
\alias{OR}
\title{Computes the odds ratio between two probabilities}
\usage{
OR(p1, p2)
}
\arguments{
\item{p1}{a probability}

\item{p2}{another probability}
}
\value{
OR=(p1/(1-p1))/(p2/(1-p2))
}
\description{
Computes the odds ratio between two probabilities
}
\examples{
OR(.5,.2)

}
