% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plotGR}
\alias{plotGR}
\title{Produces a plot of the values of the Gelman Rubin stats to determine
visually convergence (and see clearly which node has reached it)}
\usage{
plotGR(m)
}
\arguments{
\item{m}{is an object in the class jags or bugs (the output of the MCMC run)}
}
\value{
the graph with the Gelmn Rubin statistics plot
}
\description{
Produces a plot of the values of the Gelman Rubin stats to determine
visually convergence (and see clearly which node has reached it)
}
\examples{
\dontrun{
}

}
