% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.bugs}
\alias{print.bugs}
\title{Printing a bugs object}
\usage{
\method{print}{bugs}(x, digits = 3, intervals = c(0.025, 0.25, 0.5, 0.75, 0.975), ...)
}
\arguments{
\item{x}{an object of class `bugs', see \code{\link{bugs}} for details}

\item{intervals}{the quantiles for the posterior distribution to
be displayed in the summary statistics table}

\item{...}{further arguments to \code{\link{print}}}

\item{digits.summary}{rounding for tabular output on the console (default is
to round to 1 decimal place)}
}
\description{
Printing a \code{bugs} object
}
\seealso{
\code{\link{bugs}}
}
\keyword{print}
