% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.rjags}
\alias{print.rjags}
\title{Printing a jags object}
\usage{
\method{print}{rjags}(x, digits = 3, intervals = c(0.025, 0.25, 0.5, 0.75, 0.975), ...)
}
\arguments{
\item{x}{an object of class `rjags', see \code{\link{jags}} for details}

\item{digits}{rounding for tabular output on the console (default is
to round to 1 decimal place)}

\item{intervals}{the quantiles for the posterior distribution to
be displayed in the summary statistics table}

\item{...}{further arguments to \code{\link{print}}}
}
\description{
Printing a \code{jags} object
}
\seealso{
\code{\link{jags}}
}
\author{
Gianluca Baio
}
\keyword{print}
