% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stats2}
\alias{stats2}
\title{Computes and prints summary statistics for a vector or matrix of
simulated values - tidyverse style}
\usage{
stats2(x, na.rm = TRUE)
}
\arguments{
\item{x}{A vector or a matrix containing simulations from, eg, BUGS}
}
\description{
Computes and prints summary statistics for a vector or matrix of
simulated values - tidyverse style
}
\examples{
x=rnorm(1000)
stats2(x)

}
