% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{betaplot}
\alias{betaplot}
\title{Trial-and-error Beta plot}
\usage{
betaplot()
}
\description{
Provides a quick and dirty, trial-and-error tool to identify suitable values for the the
parameters of a Beta distribution to match set properties (eg mean, sd, 95\% interval)
}
\examples{
\dontrun{
}
}
\author{
Gianluca Baio
}
\keyword{Beta}
\keyword{distribution}
