% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{coefplot}
\alias{coefplot}
\title{Coefplot for the parameters in the model}
\usage{
coefplot(x, low = 0.025, upp = 0.975, parameter = NULL, ...)
}
\arguments{
\item{x}{an object of class `bugs', see \code{BUGS}, or of class
'jags', see \code{JAGS} for details}

\item{low}{the lower quantile to consider (default 2.5 percentile)}

\item{upp}{the upper quantile to consider (default 97.5 percentile)}

\item{parameter}{a vector of strings with the names of the parameters to be
included. Defaults to all those in the original model, but can be a
vector eg \code{c("par1","par2")}}

\item{...}{Additional options}
}
\description{
Creates a plot showing the mean and an interval estimate for the posterior
distributions in a given model.
}
\examples{
\dontrun{
}
}
\seealso{
\code{BUGS}, \code{JAGS}
}
\author{
Gianluca Baio
}
\keyword{Coefficient}
\keyword{plots}
