% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{diagplot}
\alias{diagplot}
\title{Specialised diagnostic plots}
\usage{
diagplot(x, what = "Rhat", label = FALSE, ...)
}
\arguments{
\item{x}{an object of class `bugs', see \code{BUGS}, or of class
'jags', see \code{JAGS} for details}

\item{what}{A string indicating what diagnostic measure should be plotted.
Options are 'Rhat' (default), indicating the PSR statistic, or 'n.eff',
indicating the effective sample size}

\item{label}{A logical input. If set to 'FALSE' (default), then does not
include text labels next to each node}

\item{...}{Additional options}
}
\description{
Creates a plot showing the output of convergence indicators, such as
the Potential Scale Reduction and the effective sample size
}
\examples{
\dontrun{
}
}
\seealso{
\code{BUGS}, \code{JAGS}
}
\author{
Gianluca Baio
}
\keyword{Diagnostic}
\keyword{plots}
