% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{gammaplot}
\alias{gammaplot}
\title{Trial-and-error Gamma plot}
\usage{
gammaplot()
}
\description{
Provides a quick and dirty, trial-and-error tool to identify suitable values for the the
parameters of a Gamma distribution to match set properties (eg mean, sd, 95\% interval)
}
\examples{
\dontrun{
}
}
\author{
Gianluca Baio
}
\keyword{Gamma}
\keyword{distribution}
