% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ilogit}
\alias{ilogit}
\title{Computes the inverse logit of a number between -infinity and +infinity}
\usage{
ilogit(x)
}
\arguments{
\item{x}{a real number}
}
\value{
inverse-logit(x) = exp(x)/(1+exp(x))
}
\description{
Computes the inverse logit of a number between -infinity and +infinity
}
\examples{
ilogit(2)

}
