% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{logitPar}
\alias{logitPar}
\title{Computes the parameters of a Gamma distribution so that the mean and
standard dev are the input (m,s)}
\usage{
logitPar(low, upp)
}
\arguments{
\item{low}{The lower extreme of an implied range that is supposed to cover
"most" of the mass under the natural scale of the distribution of the
parameter (defined in \link{0,1})}

\item{upp}{The upper extreme of an implied range that is supposed to cover
"most" of the mass under the natural scale of the distribution of the
parameter (defined in \link{0,1})}
}
\value{
The list of relevant output including the values for the
parameters of the normal distribution (mulogit and sigmalogit),
\strong{on the logit scale}
}
\description{
Computes the parameters of a Gamma distribution so that the mean and
standard dev are the input (m,s)
}
\examples{
gammaPar(12,3)

}
