% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stats}
\alias{stats}
\title{Computes and prints summary statistics for a vector or matrix of
simulated values}
\usage{
stats(x, dim = 2)
}
\arguments{
\item{x}{A vector or a matrix containing simulations from, eg, BUGS}
}
\value{
A table with some specific summary statistics
}
\description{
Computes and prints summary statistics for a vector or matrix of
simulated values
}
\examples{
x=rnorm(1000)
stats(x)

}
