% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{traceplot}
\alias{traceplot}
\title{Tidyverse based function to do traceplots}
\usage{
traceplot(x, parameter = NULL, ...)
}
\arguments{
\item{x}{an object of class `bugs', see \code{\link{bugs}}, or of class
'jags', see \code{\link{jags}} for details}

\item{parameter}{a string with the name of the parameter for which to show
the traceplot. Can be a vector, eg \code{c("par1","par2")}}

\item{...}{further arguments to \code{\link{traceplot}}}
}
\description{
Traceplot for a \code{bugs} or \code{jags} object
}
\seealso{
\code{\link{bugs}}, \code{\link{jags}}
}
\author{
Gianluca Baio
}
